The VHDL codes with seperated modules
- Catch The Fruits -
In my project, I am going to use VGA out of the FPGA board to design a game like Atari games.
The other parts of the FPGA like buttons or switches will be required. In the game, the player will try to
catch fruits falling down arbitrary locations from the top of the screen and also have a basket to catch
them. The basket located at the bottom of the screen and it can be moved left or right. There will be 3
or more different fruits but, one of them is the main target and the fruit can be change randomly so
the player should be aware of which fruit is targeted when he/she plays. The fruits can be pear, apple or
watermelon to make game lovable. Moreover, to make the game funnier, the fall speed of the fruits
will be changeable and it may be slow, normal or fast. If the player overcomes and gains more points,
the game speed will be increased. Finally, If I explain my project gameplay, it will be about catching
different types of fruits that fall down and the player can catch one type of them which demonstrated
on the top right side of the screen. The player moves the basket left or right. The point gained by catching
fruits will increase via player ability because the game speed will increase so the game will be hard.
However, the player can adjust the game speed by switches.
(Expected Screen Display)
In the project, we will need some components that provide some requirements. The
controllers component takes data from the user. Other components will enable the circuit to monitor and
design the screen elements.
The components on the FPGA board which I will use in my project are
1. Buttons of basys 3 board will be used to control the basket.
2. Pins will be to control game speed automatically,
3. Vga out of basys 3 is going to used to show the game elements on the screen.
4. The clock signal(100 MHz) of the board will be used for multiple reasons like a VGA
driver(must use the signal).
The VGA driver component use clk 100 MHz to generate synchronization of the pixels and also
give coordinates of which pixel is shown on the screen.
The controller module takes user inputs and is delivered to pixel generator module.
Pixel generator module combines the inputs from VGA driver and controller module to create
meaningful images on the screen.
A user going to try catching fruit which is displayed left top of the screen to gain points but if they
can tackle, the speed of the game will be increased automatically by a circuit or by the user. The game may
store the highest point and show it on the screen.